+++ packages/amplify-ui-components/src/components/amplify-sign-up/amplify-sign-up.tsx
@@ -14,6 +14,7 @@ import {
 import {
 	AuthState,
 	AuthStateHandler,
+	UsernameAlias,
 	UsernameAliasStrings,
 } from '../../common/types/auth-types';
 import { SignUpAttributes } from '../../common/types/auth-types';
@@ -148,7 +149,13 @@ export class AmplifySignUp {
 		}
 		try {
 			if (!this.signUpAttributes.username) {
-				throw new Error(Translations.EMPTY_USERNAME);
+				if (this.usernameAlias === UsernameAlias.email) {
+					throw new Error(Translations.EMPTY_EMAIL);
+				} else if (this.usernameAlias === UsernameAlias.phone_number) {
+					throw new Error(Translations.EMPTY_PHONE);
+				} else {
+					throw new Error(Translations.EMPTY_USERNAME);
+				}
 			}
 			if (this.signUpAttributes.username.indexOf(' ') >= 0) {
 				throw new Error(Translations.USERNAME_REMOVE_WHITESPACE);